dcetl/requirements.txt
jinja2
dask
pyyaml
cx-Oracle==6.3.1
future>=0.16.0
jieba>=0.39
numpy==1.14.3
pandas==0.25.3
psycopg2-binary==2.7.4
pymongo==3.10.1
pymssql==2.1.5
PyMySQL==0.9.2
redis==2.10.6
six==1.11.0
sklearn-pandas==1.8.0
sqlalchemy==1.2.9
SQLAlchemy-Utc==0.10.0
SQLAlchemy-Utils==0.33.2
sqlparse==0.2.4
xlrd==1.1.0
PyGreSQL==5.2.1
cloudpickle>=0.5.2
croniter
pyspark==2.4.7
